Fixed GifReader classes and ResHelper class
The GifReader classes got a NullPointerException when run because they tried to get

ResHelper.class.getResource() ...

getResource() should be called on the a ClassLoader not on the class

Now the GIfReader classes call

ResHelper.classloader ....

This is a static variable on ResHelper.
ResHelper gets it's own Classloader that can be used from other classes to simplify their code.